created utility function for responses
